API Integration 'v1/forecast'

+++ app/src/main/java/com/ray/weather/data/remote/WeatherNetworkDataSource.kt
@@ -7,6 +7,6 @@ import com.ray.weather.data.remote.model.ForecastNetworkResponse
  */
 interface WeatherNetworkDataSource {
 
-    suspend fun getCurrentWeather(lat: Long, lng: Long): ForecastNetworkResponse
+    suspend fun getCurrentWeather(lat: Double, lng: Double): ForecastNetworkResponse
 
 }
